+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * Translate `\` in lambda expression to `OP-LAMBDA`. (#18 @renkun-ken)
 
+* Drop all control characters, except horizaontal tab and newline (#19).
+
 # 1.0.4
 
 * Translate ] tokens to `OP-RIGHT-BRACKET` instead of